Fixing testing output that came from refactoring StringIO to python3 compatible modules
